chore(users): split users to calendar & events #37
