+++ backend/src/main/java/unischedule/events/controller/EventsController.java
@@ -1,9 +1,7 @@
 package unischedule.events.controller;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import unischedule.events.dto.EventModifyRequestDto;
 import unischedule.events.service.EventsService;
-import unischedule.users.dto.EventGetResponseDto;
+import unischedule.events.dto.EventGetResponseDto;
 
 @RestController
 @RequestMapping("/api/events")
