+++ backend/src/main/java/unischedule/events/dto/EventCreateResponseDto.java
@@ -1,7 +1,7 @@
-package unischedule.users.dto;
+package unischedule.events.dto;
 
 import java.time.LocalDateTime;
-import unischedule.users.entity.Event;
+import unischedule.events.entity.Event;
 
 public record EventCreateResponseDto(Long eventId, Long userId, String title,
                                      String description, LocalDateTime startTime,
