+++ backend/src/main/java/unischedule/events/repository/EventRepository.java
@@ -1,10 +1,10 @@
-package unischedule.users.repository;
+package unischedule.events.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import unischedule.users.entity.Event;
+import unischedule.events.entity.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
